Remove PaymentStruct.shop from Remittance
There is no need to independantly store the shop owner's address. This
is because the shop owner's address is 'stored' in the puzzle generated
when setting up the funds. When claiming the funds only the correct shop
owner's address will be able to generate the puzzle required to release
the funds.